[IMP] mis_builder: Avoid problem on integration CIs
Executing tests on the moment of the load, you have the risk of having an error
because a field declared as required by other module has already put the SQL constraint
on the install phase (previous to test phase), but the module is not loaded yet
for providing the default value, getting a not null constraint. An example is the field
"sale_warn" in `res.company` provided by sale module.

Marking them to be executed after all modules have been loaded, you avoid that problem.

I also rename the date range type created on tests for avoiding to collapse on
duplicated name with other modules that add the generic "Year" date type.